requirements.txt
amqp==5.1.1
backports.zoneinfo==0.2.1
billiard==4.1.0
boto3==1.18.49
botocore==1.21.65
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==2.1.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
kombu==5.3.1
MarkupSafe==2.1.3
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.5.2
six==1.16.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.16.2
